+++ cmd/root/root.go
@@ -76,6 +76,7 @@ func New(ctx context.Context) *cobra.Command {
 		// Configure our user agent with the command that's about to be executed.
 		ctx = withCommandInUserAgent(ctx, cmd)
 		ctx = withUpstreamInUserAgent(ctx)
+		ctx = withCommandExecIdInUserAgent(ctx)
 		cmd.SetContext(ctx)
 		return nil
 	}
